# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Color Rush Live",
    "version": "0.2.0"
  },
  "paths": {
...
    "/simulation/commands": {
      "post": {
        "summary": "Simulation Commands",
        "operationId": "simulation_commands_simulation_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulation Commands Simulation Commands Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulation/rounds/start": {
      "post": {
        "summary": "Simulation Start",
        "operationId": "simulation_start_simulation_rounds_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimStartRound"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulation Start Simulation Rounds Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulation/rounds/close": {
      "post": {
        "summary": "Simulation Close",
        "operationId": "simulation_close_simulation_rounds_close_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimStartRound"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulation Close Simulation Rounds Close Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulation/rounds/drain": {
      "post": {
        "summary": "Simulation Drain",
        "operationId": "simulation_drain_simulation_rounds_drain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimStartRound"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulation Drain Simulation Rounds Drain Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulation/rounds/spin": {
      "post": {
        "summary": "Simulation Spin",
        "operationId": "simulation_spin_simulation_rounds_spin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimSpin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulation Spin Simulation Rounds Spin Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulation/rounds/settle": {
      "post": {
        "summary": "Simulation Settle",
        "operationId": "simulation_settle_simulation_rounds_settle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimStartRound"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulation Settle Simulation Rounds Settle Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Color": {
        "type": "string",
        "enum": [
          "red",
          "gold",
          "green"
        ],
        "title": "Color"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SimCommand": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "broadcast_id": {
            "type": "string",
            "title": "Broadcast Id",
            "default": "sim-broadcast"
          },
          "provider_channel_id": {
            "type": "string",
            "title": "Provider Channel Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "published_at": {
            "type": "string",
            "title": "Published At"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "provider_channel_id",
          "display_name",
          "text",
          "published_at",
          "message_id"
        ],
        "title": "SimCommand"
      },
      "SimSpin": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "fence_token": {
            "type": "integer",
            "title": "Fence Token"
          },
          "forced": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Color"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "fence_token"
        ],
        "title": "SimSpin"
      },
      "SimStartRound": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "fence_token": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Fence Token"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "fence_token"
        ],
        "title": "SimStartRound"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}